+++ lodg/src/headers/gather.h
@@ -2,6 +2,6 @@
 #define GLOBS_H
 
 struct monitoring_data_entry* gather_metric(struct monitoring_point* point);
-int get_metric_value(struct monitoring_point* point);
+float get_metric_value(struct monitoring_point* point);
 
 #endif
